RUNTIME: Code rearranged between Memory.* and Types.*
